--- src/systems/ec/implementations/react/templates/pages/call-tender/index.js
@@ -1,1 +1,0 @@
-export default from './src/CallTenderPage';
