+++ src/systems/ec/implementations/react/templates/pages/call-tenders/index.js
@@ -1,0 +1,1 @@
+export default from './src/CallTendersPage';
