+++ src/systems/ec/implementations/react/templates/pages/call-tenders/package.json
@@ -1,6 +1,6 @@
 {
   "private": true,
-  "name": "@ecl/ec-react-call-tender-page",
+  "name": "@ecl/ec-react-call-tenders-page",
   "author": "European Commission",
   "license": "EUPL-1.1",
   "version": "2.15.0",
@@ -29,7 +29,7 @@
   },
   "devDependencies": {
     "@ecl/ec-layout-grid": "^2.15.0",
-    "@ecl/ec-specs-call-tender-page": "^2.15.0",
+    "@ecl/ec-specs-call-tenders-page": "^2.15.0",
     "@ecl/ec-utility-spacing": "^2.15.0",
     "@ecl/ec-utility-typography": "^2.15.0"
   },
