+++ src/systems/ec/specs/templates/call-tenders-page/package.json
@@ -1,9 +1,9 @@
 {
-  "name": "@ecl/ec-specs-call-tender-page",
+  "name": "@ecl/ec-specs-call-tenders-page",
   "author": "European Commission",
   "license": "EUPL-1.1",
   "version": "2.15.0",
-  "description": "ECL EC Call For Tender Page Specs",
+  "description": "ECL EC Call For Tenders Page Specs",
   "publishConfig": {
     "access": "public"
   },
