+++ src/website/src/pages/ec/content-types/call/tenders/docs/code.mdx
@@ -4,13 +4,13 @@ order: 2
 ---
 
 import Playground from '@ecl/website-components/Playground';
-import CallTenderPageExample from '@ecl/ec-react-call-tender-page/examples/Default';
+import CallTendersPageExample from '@ecl/ec-react-call-tenders-page/examples/Default';
 
 <Playground
   system="ec"
   selectedKind="templates-pages"
-  selectedStory="call-for-tender"
+  selectedStory="call-for-tenders"
   hideDemo
 >
-  <CallTenderPageExample />
+  <CallTendersPageExample />
 </Playground>
