+++ src/website/src/pages/ec/content-types/call/tenders/index.md
@@ -4,5 +4,5 @@ defaultTab: usage
 status: ready
 playground:
   system: ec
-  path: /story/templates-pages--call-for-tender
+  path: /story/templates-pages--call-for-tenders
 ---
